CBL-164: Move most of CouchbaseLite to CouchbaseLiteInternal hiding CouchbaseLite.getExecutionService() Add CBL debugging flag
